Add support for floating point numbers

+++ calc.py
@@ -4,9 +4,9 @@ def compute(input_string):
     Example: '5 + 7' -> 12
     """
     values = input_string.split(' ')
-    num0 = int(values[0])
+    num0 = float(values[0])
     operator = values[1]
-    num1 = int(values[2])
+    num1 = float(values[2])
     
     if operator == '+':
         return num0 + num1
